Combine RunConfig and Run
We want to make runs more easily usable. Making changes is easier when only one class needs to be modified. Also makes it easier to learn and use the framework.
